square-model-inference-api/inference_server/requirements.txt
uvicorn
fastapi
pydantic
loguru
python-dotenv
adapter-transformers
sentencepiece
torch